chore: move states to composables
